Refactor UsersController and Add new API to update existing user accounts
